backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
passlib[bcrypt]
python-jose[cryptography]
